[24.03.24] 31230 풀이

+++ bronze/B_15964.java
@@ -8,8 +8,8 @@ public class B_15964 {
   public static void main(String[] args) throws Exception {
     BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
     StringTokenizer st = new StringTokenizer(br.readLine());
-    int A = Integer.parseInt(st.nextToken());
-    int B = Integer.parseInt(st.nextToken());
+    long A = Integer.parseInt(st.nextToken());
+    long B = Integer.parseInt(st.nextToken());
 
     System.out.println((A + B) * (A - B));
   }
